code cleanup: get rid of obsolete aggregate transformers, separate package for aggregate encoders
