+++ src/main/java/at/ac/tuwien/kr/alpha/grounder/transformation/aggregates/encoders/ElementRuleDelegatingEncoder.java
@@ -1,4 +1,4 @@
-package at.ac.tuwien.kr.alpha.grounder.transformation.aggregates;
+package at.ac.tuwien.kr.alpha.grounder.transformation.aggregates.encoders;
 
 import java.util.Collections;
 
@@ -7,6 +7,7 @@ import at.ac.tuwien.kr.alpha.common.atoms.Atom;
 import at.ac.tuwien.kr.alpha.common.program.InputProgram;
 import at.ac.tuwien.kr.alpha.common.rule.BasicRule;
 import at.ac.tuwien.kr.alpha.common.rule.head.NormalHead;
+import at.ac.tuwien.kr.alpha.grounder.transformation.aggregates.AggregateRewritingContext;
 import at.ac.tuwien.kr.alpha.grounder.transformation.aggregates.AggregateRewritingContext.AggregateInfo;
 
 // TODO do this via decorator pattern on top of any encoder!
