+++ src/main/java/at/ac/tuwien/kr/alpha/grounder/transformation/aggregates/encoders/SumEqualsAggregateEncoder.java
@@ -1,4 +1,4 @@
-package at.ac.tuwien.kr.alpha.grounder.transformation.aggregates;
+package at.ac.tuwien.kr.alpha.grounder.transformation.aggregates.encoders;
 
 import org.apache.commons.collections4.SetUtils;
 import org.stringtemplate.v4.ST;
@@ -22,6 +22,7 @@ import at.ac.tuwien.kr.alpha.common.terms.VariableTerm;
 import at.ac.tuwien.kr.alpha.grounder.parser.ProgramParser;
 import at.ac.tuwien.kr.alpha.grounder.transformation.EnumerationRewriting;
 import at.ac.tuwien.kr.alpha.grounder.transformation.PredicateInternalizer;
+import at.ac.tuwien.kr.alpha.grounder.transformation.aggregates.AggregateRewritingContext;
 import at.ac.tuwien.kr.alpha.grounder.transformation.aggregates.AggregateRewritingContext.AggregateInfo;
 
 public class SumEqualsAggregateEncoder extends AbstractAggregateEncoder {
